+++ src/components/AddExpensePage.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+const AddExpensePage = () => (
+  <div>
+    This is from my add expense component
+  </div>
+);
+
+export default AddExpensePage;
