+++ src/components/EditExpensePage.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+const EditExpensePage = () => (
+  <div>
+    This is from my edit expense component
+  </div>
+);
+
+export default EditExpensePage;
