+++ src/components/ExpenseDashboardPage.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+const ExpenseDashboardPage = () => (
+  <div>
+    This is from my dashboard component
+  </div>
+);
+
+export default ExpenseDashboardPage;
