+++ src/components/Header.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import {NavLink} from 'react-router-dom';
+
+const Header = () => (
+  <header>
+    <h1>Expensify</h1>
+    <NavLink to="/" activeClassName="is-active" exact={true} >Dashboard</NavLink>
+    <NavLink to="/create" activeClassName="is-active" >Add Expense</NavLink>
+    <NavLink to="/edit" activeClassName="is-active" >Edit Expense</NavLink>
+    <NavLink to="/help" activeClassName="is-active" >Help</NavLink>
+  </header>
+);
+
+export default Header;
