+++ src/components/HelpPage.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+const HelpPage = () => (
+  <div>
+    This is from my help page component
+  </div>
+);
+
+export default HelpPage;
